created BB.AudioBase and now extending appropriate Audio modules from it
