+++ examples/basic/audio-fx-filter/main.js
@@ -44,7 +44,7 @@ function updateWaveform( buffer ){
 	filter.type = filt.node.type;
 	filter.frequency.value = filt.frequency;
 	filter.Q.value = filt.Q;
-	filter.gain.value = filt.gain;
+	filter.gain.value = filt.fgain;
 	source.connect(filter);
 	filter.connect(offlineContext.destination);
 	source.start(0);
@@ -227,7 +227,7 @@ dat_gui_lib.onload = function(){
 
     gui.add( filt.node, 'type', types ).onChange( drawFilterShape );
     gui.add( filt, 'frequency', 20, 5000 ).onChange( drawFilterShape );
-    gui.add( filt, 'gain', -10, 10 ).onChange( drawFilterShape );
+    gui.add( filt, 'fgain', -10, 10 ).onChange( drawFilterShape );
     gui.add( filt, 'Q', 0, 20 ).onChange( drawFilterShape );
     gui.add( set, "update_waveform");
