+++ src/BB.AudioNoise.js
@@ -2,9 +2,10 @@
 /**
  * A module for creating differnt kinds of noise ( defined as any nonconventional and/or mathematically calculated sound buffer ) 
  * @module BB.AudioNoise
+ * @extends BB.AudioBase
  */
-define(['./BB', './BB.Audio', './BB.Detect' ],
-function(  BB,        Audio,        Detect ){
+define(['./BB', './BB.AudioBase' ],
+function(  BB,        AudioBase ){
 
 	'use strict';
 
@@ -12,6 +13,7 @@ function(  BB,        Audio,        Detect ){
      * A module for creating differnt kinds of noise ( defined as any nonconventional and/or mathematically calculated sound buffer ) 
      * @class BB.AudioNoise
      * @constructor
+     * @extends BB.AudioBase
      * 
      * @param {Object} config this can either be a string (type of noise "white","pink","brown","perlin" ), or a custom function ( see example further down ) or a config object to initialize the Noise
      * can contain the following:<br><br>
@@ -64,6 +66,8 @@ function(  BB,        Audio,        Detect ){
 	
 	BB.AudioNoise = function( config ) {
 
+		BB.AudioBase.call(this, config);
+
 		var types = ["white","pink","brown","perlin"];
 
 		if(typeof config === "undefined"){ config = {}; }
@@ -74,38 +78,7 @@ function(  BB,        Audio,        Detect ){
 			throw new Error('BB.AudioNoise: type should be either "white","pink","brown","perlin" or be a custom function');
 		}	
 
-		// the AudioContext to be used by this module 
-		if( typeof BB.Audio.context === "undefined" )
-			throw new Error('BB Audio Modules require that you first create an AudioContext: BB.Audio.init()');
-
-		if( BB.Audio.context instanceof Array ){
-			if( typeof config === "undefined" || typeof config.context === "undefined" )
-				throw new Error('BB.AudioNoise: BB.Audio.context is an Array, specify which { context:BB.Audio.context[?] }');
-			else 
-				this.ctx = config.context;		
-		} else {
-			this.ctx = BB.Audio.context;
-		}	
-
-		//
 		
-		// default destination is context destination
-		// unless otherwise specified in { connect:AudioNode }
-		this.gain       = this.ctx.createGain();    
-		if( typeof config.connect !== 'undefined' ){
-			if( config.connect instanceof AudioDestinationNode || config.connect instanceof AudioNode ) 
-				this.gain.connect( config.connect );
-			else 
-				throw new Error('BB.AudioNoise: connect property expecting an AudioNode');
-		
-		} else {
-			this.gain.connect( this.ctx.destination );
-		}
-
-		this._volume = (typeof config.volume !== "undefined") ? config.volume : 1; // MASTER NOISE VOLUME ESSENTIALY 
-		this.gain.gain.value = this._volume;
-		this.gain.gain.name = "master";
-
 
 		this._type 		= (typeof config.type !== "undefined") ? config.type : "white";
 		if(typeof this._type==="string" && types.indexOf( this._type ) < 0 ) 
@@ -155,25 +128,9 @@ function(  BB,        Audio,        Detect ){
 
 	};
 
+ 	BB.AudioNoise.prototype = Object.create(BB.AudioBase.prototype);
+    BB.AudioNoise.prototype.constructor = BB.AudioNoise;
 
-	/**
-	* the master volume (gain node)
-	* @property volume
-	* @type Number
-	* @default 1
-	*/   
-	Object.defineProperty(BB.AudioNoise.prototype, "volume", {
-		get: function() {
-			return this._volume;
-		},
-		set: function(v) {
-			if( typeof v !== 'number'){
-				throw new Error("BB.AudioNoise.volume: expecing a number");
-			} else {
-				this.setGain( v );
-			}
-		}
-	});
 
 	/**
 	* "white","pink","brown" or "custom"
@@ -220,130 +177,6 @@ function(  BB,        Audio,        Detect ){
 			}
 		}
 	});
-
-	/**
-	* connects the Noise to a particular AudioNode or AudioDestinationNode
-	* @method connect
-	* @param  {AudioNode} destination the AudioNode or AudioDestinationNode to connect to
-	* @param  {Number} output      which output of the the Noise do you want to connect to the destination
-	* @param  {Number} input       which input of the destination you want to connect the Noise to
-	* @example  
-	* <code class="code prettyprint">  
-	*  &nbsp;BB.Audio.init();<br>
-	*  <br>
-	*  &nbsp;var whitenoise = new BB.AudioNoise({<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75,<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;type: "white"<br>
-	*  &nbsp;});<br>
-	*  <br>
-	*  &nbsp; whitenoise.connect( exampleNode );<br>
-	*  &nbsp; // connected to both default BB.Audio.context && exampleNode<br>
-	*  &nbsp; // so if exampleNode is also connected to BB.Audio.context by default,<br>
-	*  &nbsp; // ...then you've got whitenoise connected to BB.Audio.context twice<br>
-	* </code>
-	* <br>
-	* ...which looks like this ( where the first Gain is the Noise and the second is the exampleNode )<br>
-	* <img src="../assets/images/audiosampler3.png">
-	*/
-	BB.AudioNoise.prototype.connect = function( destination, output, input ){
-		if( !(destination instanceof AudioDestinationNode || destination instanceof AudioNode) )
-			throw new Error('BB.AudioNoise.connect: destination should be an instanceof AudioDestinationNode or AudioNode');
-		if( typeof output !== "undefined" && typeof output !== "number" )
-			throw new Error('BB.AudioNoise.connect: output should be a number');
-		if( typeof intput !== "undefined" && typeof input !== "number" )
-			throw new Error('BB.AudioNoise.connect: input should be a number');
-
-		if( typeof intput !== "undefined" ) this.gain.connect( destination, output, input );
-		else if( typeof output !== "undefined" ) this.gain.connect( destination, output );
-		else this.gain.connect( destination );
-	};
-
-	/**
-	* diconnects the Noise from the node it's connected to
-	* @method disconnect
-	* @param  {AudioNode} destination what it's connected to
-	* @param  {Number} output      the particular output number
-	* @param  {Number} input       the particular input number
-	* <code class="code prettyprint">  
-	*  &nbsp;BB.Audio.init();<br>
-	*  <br>
-	*  &nbsp;var whitenoise = new BB.AudioNoise({<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75,<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;type: "white"<br>
-	*  &nbsp;});<br>
-	*  <br>
-	*  &nbsp;whitenoise.disconnect(); // disconnected from default BB.Audio.context<br>
-	*  &nbsp;whitenoise.connect( exampleNode ); // connected to exampleNode only<br>
-	*  &nbsp;whitenoise.makeNote('A');<br>
-	* </code>
-	* <br>
-	* ...which looks like this ( where the first Gain is the Noise and the second is the exampleNode )<br>
-	* <img src="../assets/images/audiosampler4.png">
-	*/
-	BB.AudioNoise.prototype.disconnect = function(destination, output, input ){
-		if( typeof destination !== "undefined" &&
-			!(destination instanceof AudioDestinationNode || destination instanceof AudioNode) )
-				throw new Error('BB.AudioNoise.disconnect: destination should be an instanceof AudioDestinationNode or AudioNode');
-		if( typeof output !== "undefined" && typeof output !== "number" )
-			throw new Error('BB.AudioNoise.disconnect: output should be a number');
-		if( typeof input !== "undefined" && typeof input !== "number" )
-			throw new Error('BB.AudioNoise.disconnect: input should be a number');
-
-		if( typeof input !== "undefined" ) this.gain.disconnect( destination, output, input );
-		else if( typeof output !== "undefined" ) this.gain.disconnect( destination, output );
-		else if( typeof destination !== "undefined" ) this.gain.disconnect( destination );
-		else  this.gain.disconnect();
-	};
-
-
-	/**
-	* sets the gain level of the Noise ( in a sense, master volume control )
-	* @method setGain
-	* @param {Number} num a float value, 1 being the default volume, below 1 decreses the volume, above one pushes the gain
-	* @param {Number} ramp value in seconds for how quickly/slowly to ramp to the new value (num) specified
-	*
-	* @example  
-	* <code class="code prettyprint">  
-	*  &nbsp;BB.Audio.init();<br>
-	*  <br>
-	*  &nbsp;var whitenoise = new BB.AudioNoise({<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;type: "white",<br>
-	*  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75<br>
-	*  &nbsp;});<br>
-	*  <br>
-	*  &nbsp; whitenoise.setGain( 0.25, 2 ); // lower's volume from 0.75 to 0.25 in 2 seconds <br>
-	*  // if no ramp value is needed, you could alternatively do<br>
-	*  &nbsp; whitenoise.volume = 0.5; // immediately jumps from 0.25 to 0.5 <br>
-	* </code>
-	*/
-	BB.AudioNoise.prototype.setGain = function( num, gradually ){
-		if( typeof num !== "number" )
-			throw new Error('BB.AudioNoise.setGain: first argument expecting a number');
-		
-		this._volume = num;
-
-		if(typeof gradually !== "undefined"){
-			if( typeof num !== "number" )
-				throw new Error('BB.AudioNoise.setGain: second argument expecting a number');
-			else
-				this._globalGainUpdate( gradually );
-		}
-		else { this._globalGainUpdate(0); }
-	};
-
-	BB.AudioNoise.prototype._sec2tc = function( sec ){
-		return ( sec * 2 ) / 10;
-	};
-
-	BB.AudioNoise.prototype._globalGainUpdate = function( gradually ){
-		if(typeof gradually === "undefined") gradually = 0;    
-
-		if(BB.Detect.browserInfo.name=="Firefox"){
-			this.gain.gain.setTargetAtTimePoly( this._volume, 0, gradually);
-		} else {
-			this.gain.gain.setTargetAtTime( this._volume, 0, this._sec2tc(gradually));
-		}    
-	};
 
 	// ---------------
 
@@ -526,56 +359,6 @@ function(  BB,        Audio,        Detect ){
 		this.isPlaying = false;
 		if(typeof fromMakeNoise==="undefined" || fromMakeNoise===false) this._isOn = false;
 	};
-
-
-	// ... polyfills .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... ....
-
-	if(BB.Detect.browserInfo.name=="Firefox"){   
-		// !!! *** !!! *** !!! *** !!! *** !!! *** !!! *** !!! 
-		// !!!  THIS POLY FILL IS STILL UNDERCONSTRUCTION  !!!
-		// !!! *** !!! *** !!! *** !!! *** !!! *** !!! *** !!! 
-		// 
-		AudioParam.prototype.setTargetAtTimePoly = function( target, startTime, seconds ){
-
-		    // this.value = target;
-
-		    var polyTimeout = null;
-		    var polyInterval = null;
-		    startTime = startTime * 1000;
-		    var self = this;
-		    
-		    // if(typeof polyTimeout !== "undefined"){ clearTimeout( polyTimeout ); polyTimeout = undefined; }  
-		    
-		    polyTimeout = setTimeout(function(){
-		        
-		        
-		        if(seconds===0){     
-		            self.value = target;
-		            // console.log(self.name,'jump to '+target+", now "+self.value)             
-		        } else {
-		            // if(typeof polyInterval !== "undefined"){ clearInterval( polyInterval ); polyInterval = undefined; }  
-
-		            var delta =  target - self.value;
-		            var inc = delta / (seconds*60);
-		            var dir = (delta>=0) ? "up" : "down";
-		            
-		            polyInterval = setInterval(function(){                      
-		                
-		                self.value += inc*2; // why 2? well b/c for some f'n reason it goes too slow otherwise
-
-		                if(dir=="up"){
-		                    if( self.value >= target){clearInterval( polyInterval ); }
-		                } else {
-		                    if( self.value <= target){clearInterval( polyInterval ); }
-		                }
-		                
-		            }, 1000/60 );
-		        }
-
-		    }, startTime );
-		};
-
-	}
 
 	return BB.AudioNoise;
 });
