+++ src/BB.AudioTone.js
@@ -1,9 +1,10 @@
 /**
  * A module for creating oscillating tones ( periodic wave forms ) along with music theory utilities
  * @module BB.AudioTone
+ * @extends BB.AudioBase
  */
-define(['./BB', './BB.Audio', './BB.Detect' ],
-function(  BB,        Audio,        Detect ) {
+define(['./BB', './BB.AudioBase' ],
+function(  BB,        AudioBase ) {
 
     'use strict';
     
@@ -11,6 +12,7 @@ function(  BB,        Audio,        Detect ) {
      * A module for creating oscillating tones ( periodic wave forms ) along with music theory utilities
      * @class BB.AudioTone
      * @constructor
+     * @extends BB.AudioBase
      * 
      * @param {Object} config A config object to initialize the Tone,
      * can contain the following:<br><br>
@@ -59,39 +61,11 @@ function(  BB,        Audio,        Detect ) {
 
     BB.AudioTone = function( config ) {
 
-        // the AudioContext to be used by this module 
-        if( typeof BB.Audio.context === "undefined" )
-            throw new Error('BB Audio Modules require that you first create an AudioContext: BB.Audio.init()');
-        
-        if( BB.Audio.context instanceof Array ){
-            if( typeof config === "undefined" || typeof config.context === "undefined" )
-                throw new Error('BB.AudioTone: BB.Audio.context is an Array, specify which { context:BB.Audio.context[?] }');
-            else {
-                this.ctx = config.context;
-            }
-        } else {
-            this.ctx = BB.Audio.context;
-        }
+        BB.AudioBase.call(this, config);
+
 
         if( !config ) config = {}; // instead of below? ...bad practice?
         // if( !config ) throw new Error('BB.AudioTone: requires a config object');
-
-        // default destination is context destination
-        // unless otherwise specified in { connect:AudioNode }
-        this.gain       = this.ctx.createGain();    
-        if( typeof config.connect !== 'undefined' ){
-            if( config.connect instanceof AudioDestinationNode ||
-                config.connect instanceof AudioNode ) 
-                this.gain.connect( config.connect );
-            else {
-                throw new Error('BB.AudioTone: connect property expecting an AudioNode');
-            }
-        } else {
-            this.gain.connect( this.ctx.destination );
-        }
-        this._volume = (typeof config.volume !== "undefined") ? config.volume : 1; // MASTER OSC VOLUME ESSENTIALY 
-        this.gain.gain.value = this._volume;
-        this.gain.gain.name = "master";
 
          /**
          * the type of wave
@@ -223,6 +197,8 @@ function(  BB,        Audio,        Detect ) {
         };
     };
 
+    BB.AudioTone.prototype = Object.create(BB.AudioBase.prototype);
+    BB.AudioTone.prototype.constructor = BB.AudioTone;
 
     /**
      * the wave Array/Object
@@ -264,135 +240,6 @@ function(  BB,        Audio,        Detect ) {
             }
         }
     });
-
-
-    /**
-     * the master volume
-     * @property volume
-     * @type Number
-     * @default 1
-     */   
-    Object.defineProperty(BB.AudioTone.prototype, "volume", {
-        get: function() {
-            return this._volume;
-        },
-        set: function(v) {
-            if( typeof v !== 'number'){
-                throw new Error("BB.AudioTone.volume: expecing a number");
-            } else {
-                this.setGain( v );
-            }
-        }
-    });
-
-    /**
-     * connects the Tone to a particular AudioNode or AudioDestinationNode
-     * @method connect
-     * @param  {AudioNode} destination the AudioNode or AudioDestinationNode to connect to
-     * @param  {Number} output      which output of the the Tone do you want to connect to the destination
-     * @param  {Number} input       which input of the destination you want to connect the Tone to
-     * @example  
-     * <code class="code prettyprint">  
-     *  &nbsp;BB.Audio.init();<br>
-     *  <br>
-     *  &nbsp;var O = new BB.AudioTone({<br>
-     *  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75,<br>
-     *  &nbsp;&nbsp;&nbsp;&nbsp;type: "square"<br>
-     *  &nbsp;});<br>
-     *  <br>
-     *  &nbsp; O.connect( exampleNode );<br>
-     *  &nbsp; // connected to both default BB.Audio.context && exampleNode<br>
-     *  &nbsp; // so if exampleNode is also connected to BB.Audio.context by default,<br>
-     *  &nbsp; // ...then you've got O connected to BB.Audio.context twice<br>
-     * </code>
-     * <br>
-     * ...which looks like this ( where the first Gain is the Tone and the second is the exampleNode )<br>
-     * <img src="../assets/images/audiosampler3.png">
-     */
-    BB.AudioTone.prototype.connect = function( destination, output, input ){
-        if( !(destination instanceof AudioDestinationNode || destination instanceof AudioNode) )
-            throw new Error('BB.AudioTone.connect: destination should be an instanceof AudioDestinationNode or AudioNode');
-        if( typeof output !== "undefined" && typeof output !== "number" )
-            throw new Error('BB.AudioTone.connect: output should be a number');
-        if( typeof intput !== "undefined" && typeof input !== "number" )
-            throw new Error('BB.AudioTone.connect: input should be a number');
-
-        if( typeof intput !== "undefined" ) this.gain.connect( destination, output, input );
-        else if( typeof output !== "undefined" ) this.gain.connect( destination, output );
-        else this.gain.connect( destination );
-
-    };
-
-    /**
-     * diconnects the Tone from the node it's connected to
-     * @method disconnect
-     * @param  {AudioNode} destination what it's connected to
-     * @param  {Number} output      the particular output number
-     * @param  {Number} input       the particular input number
-     * <code class="code prettyprint">  
-     *  &nbsp;BB.Audio.init();<br>
-     *  <br>
-     *  &nbsp;var O = new BB.AudioTone({<br>
-     *  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75,<br>
-     *  &nbsp;&nbsp;&nbsp;&nbsp;type: "square"<br>
-     *  &nbsp;});<br>
-     *  <br>
-     *  &nbsp;O.disconnect(); // disconnected from default BB.Audio.context<br>
-     *  &nbsp;O.connect( exampleNode ); // connected to exampleNode only<br>
-     *  &nbsp;O.makeNote('A');<br>
-     * </code>
-     * <br>
-     * ...which looks like this ( where the first Gain is the Tone and the second is the exampleNode )<br>
-     * <img src="../assets/images/audiosampler4.png">
-     */
-    BB.AudioTone.prototype.disconnect = function(destination, output, input ){
-        if( typeof destination !== "undefined" &&
-            !(destination instanceof AudioDestinationNode || destination instanceof AudioNode) )
-            throw new Error('BB.AudioTone.disconnect: destination should be an instanceof AudioDestinationNode or AudioNode');
-        if( typeof output !== "undefined" && typeof output !== "number" )
-            throw new Error('BB.AudioTone.disconnect: output should be a number');
-        if( typeof input !== "undefined" && typeof input !== "number" )
-            throw new Error('BB.AudioTone.disconnect: input should be a number');
-
-        if( typeof input !== "undefined" ) this.gain.disconnect( destination, output, input );
-        else if( typeof output !== "undefined" ) this.gain.disconnect( destination, output );
-        else if( typeof destination !== "undefined" ) this.gain.disconnect( destination );
-        else  this.gain.disconnect();
-    };
-
-
-    /**
-     * sets the gain level of the Oscillator ( in a sense, master volume control )
-     * @method setGain
-     * @param {Number} num a float value, 1 being the default volume, below 1 decreses the volume, above one pushes the gain
-     * @param {Number} ramp value in seconds for how quickly/slowly to ramp to the new value (num) specified
-     *
-     * @example  
-     * <code class="code prettyprint">  
-     *  &nbsp;BB.Audio.init();<br>
-     *  <br>
-     *  &nbsp;var O = new BB.AudioTone({<br>
-     *  &nbsp;&nbsp;&nbsp;&nbsp;volume: 0.75,<br>
-     *  &nbsp;});<br>
-     *  <br>
-     *  &nbsp; O.setGain( 0.25, 2 ); // lower's volume from 0.75 to 0.25 in 2 seconds <br>
-     *  // if no ramp value is needed, you could alternatively do<br>
-     *  &nbsp; O.volume = 0.5; // immediately jumps from 0.25 to 0.5 <br>
-     * </code>
-     */
-    BB.AudioTone.prototype.setGain = function( num, gradually ){
-        if( typeof num !== "number" )
-            throw new Error('BB.AudioTone.setGain: first argument expecting a number');
-        this._volume = num;
-        
-        if(typeof gradually !== "undefined"){
-            if( typeof num !== "number" )
-                throw new Error('BB.AudioTone.setGain: second argument expecting a number');
-            else
-                this._globalGainUpdate( gradually );
-        }
-        else { this._globalGainUpdate(0); }
-    };
 
 
     // ... public utils .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... ....
@@ -624,12 +471,8 @@ function(  BB,        Audio,        Detect ) {
 
     // ... private utils .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... ....
 
-
-    BB.AudioTone.prototype._sec2tc = function( sec ){
-        return ( sec * 2 ) / 10;
-    };
-
-
+    // OVERRIDE the BB.AudioBase._globalGainUpdate()
+    // 
     BB.AudioTone.prototype._globalGainUpdate = function( gradually, fromNoteOff ){
         if(typeof gradually === "undefined") gradually = 0;
         
@@ -1059,60 +902,6 @@ function(  BB,        Audio,        Detect ) {
             });
         }
     };
-
-
-
-    // ... polyfills .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... .... ....
-
-
-    if(BB.Detect.browserInfo.name=="Firefox"){   
-
-        // !!! *** !!! *** !!! *** !!! *** !!! *** !!! *** !!! 
-        // !!!  THIS POLY FILL IS STILL UNDERCONSTRUCTION  !!!
-        // !!! *** !!! *** !!! *** !!! *** !!! *** !!! *** !!! 
-        // 
-        AudioParam.prototype.setTargetAtTimePoly = function( target, startTime, seconds ){
-
-            // this.value = target;
-
-            var polyTimeout = null;
-            var polyInterval = null;
-            startTime = startTime * 1000;
-            var self = this;
-            
-            // if(typeof polyTimeout !== "undefined"){ clearTimeout( polyTimeout ); polyTimeout = undefined; }  
-            
-            polyTimeout = setTimeout(function(){
-                
-                
-                if(seconds===0){     
-                    self.value = target;
-                    // console.log(self.name,'jump to '+target+", now "+self.value)             
-                } else {
-                    // if(typeof polyInterval !== "undefined"){ clearInterval( polyInterval ); polyInterval = undefined; }  
-
-                    var delta =  target - self.value;
-                    var inc = delta / (seconds*60);
-                    var dir = (delta>=0) ? "up" : "down";
-                    
-                    polyInterval = setInterval(function(){                      
-                        
-                        self.value += inc*2; // why 2? well b/c for some f'n reason it goes too slow otherwise
-
-                        if(dir=="up"){
-                            if( self.value >= target){clearInterval( polyInterval ); }
-                        } else {
-                            if( self.value <= target){clearInterval( polyInterval ); }
-                        }
-                        
-                    }, 1000/60 );
-                }
-
-            }, startTime );
-        };
-
-    }
-
 
     return BB.AudioTone;
 });
